git-svn-id: http://svn.code.sf.net/p/distributedcl/code/trunk@200 cb344d40-e590-48b6-843f-7a98a4fc1ff3

+++ src/libdistributedcl/network/tcp_transport.h
@@ -135,7 +135,7 @@ public:
         {
             accept_thread_ptr_ = new boost::thread( &dcl::network::platform::tcp_transport< OBSERVER >::accept_loop, this );
 
-            boost::this_thread::sleep( boost::posix_time::milliseconds( 10 ) );
+            boost::this_thread::sleep( boost::posix_time::milliseconds( static_cast<int64_t>( 10 ) ) );
         }
     }
